Add user edit screen for admin

+++ frontend/src/screens/UserListScreen.js
@@ -26,7 +26,9 @@ const UserListScreen = ({ history }) => {
     }
   }, [dispatch, history, userInfo, successDelete]);
   const deleteHandler = (id) => {
-    dispatch(deleteUser(id));
+    if (window.confirm("Are you sure you want to delete this user?")) {
+      dispatch(deleteUser(id));
+    }
   };
 
   return (
@@ -61,7 +63,7 @@ const UserListScreen = ({ history }) => {
                   )}
                 </td>
                 <td>
-                  <LinkContainer to={`/admin/user/${user._id}`}>
+                  <LinkContainer to={`/admin/user/${user._id}/edit`}>
                     <Button variant="light" className="btn-sm">
                       <i className="fas fa-edit" />
                     </Button>
